+++ ar-app/app/components/App.vue
@@ -1,32 +1,33 @@
 <template>
-    <Page>
-        <ActionBar title="Welcome to NativeScript-Vue!"/>
-        <GridLayout columns="*" rows="*">
-            <Label class="message" :text="msg" col="0" row="0"/>
-        </GridLayout>
-    </Page>
+  <Page>
+    <bookmarks></bookmarks>
+  </Page>
 </template>
 
 <script>
-  export default {
-    data() {
-      return {
-        msg: 'Hello World!'
-      }
-    }
+import bookmarks from "./BookMarks";
+export default {
+  data() {
+    return {
+      msg: "Hello World!"
+    };
+  },
+  components: {
+    bookmarks
   }
+};
 </script>
 
 <style scoped>
-    ActionBar {
-        background-color: #53ba82;
-        color: #ffffff;
-    }
+ActionBar {
+  background-color: #53ba82;
+  color: #ffffff;
+}
 
-    .message {
-        vertical-align: center;
-        text-align: center;
-        font-size: 20;
-        color: #333333;
-    }
+.message {
+  vertical-align: center;
+  text-align: center;
+  font-size: 20;
+  color: #333333;
+}
 </style>
